Fix format string signedness warnings
GCC 5.0, with its new -Wformat-signedness, warns about the sign being
different between a type and the format string in printf-format
messages. Fix the sign of the format strings and types so that they
match.

https://bugs.freedesktop.org/show_bug.cgi?id=89070

+++ src/storagedata.c
@@ -303,7 +303,7 @@ storaged_ata_send_command_sync (gint                       fd,
     {
       g_set_error (error, G_IO_ERROR, G_IO_ERROR_FAILED,
                    "ATA command failed: error=0x%02x count=0x%02x status=0x%02x",
-                   output->error, output->count, output->status);
+                   (guint) output->error, (guint) output->count, (guint) output->status);
       goto out;
     }
